+++ src/tests/InfoResultTests/ResultTests.cs
@@ -20,7 +20,6 @@ using System;
 using System.Linq;
 using AggregatedGenericResultMessage;
 using AggregatedGenericResultMessage.Enums;
-using AggregatedGenericResultMessage.Extensions.Result.Messages;
 using AggregatedGenericResultMessage.Helpers;
 using InfoResultTests.Dtos;
 using InfoResultTests.Services;